refactor(core): extract ChartRenderer from Chart
Move all render pipeline logic (draw, scheduleDraw, prepareFrameData,
renderPanes, renderXAxis, clearAllCanvases, calcKLinePositions) and
their associated state (raf, cachedDrawFrame, markerManager, drawingStore,
settings, overlayHadCrosshair, xAxisCtx) into a new ChartRenderer class.

Chart keeps pluginHost, rendererPluginManager, sharedWebGLSurface
(shared with ChartIndicatorManager) and delegates render methods via
a minimal RendererDependencies interface (13 callbacks).

Also remove dead FrameData type and stale calcKWidthPx re-export.
